moved preprocessSource to its own file
